feat: Integrate custom GeoJSON markers and popups in Dash Leaflet
- Implement a custom JavaScript function for rendering markers with a specific icon for oil derrick locations.
- Define popups for each marker to display detailed information about oil derricks, including operator name, API number, well type, lease name, start date, and well status.
- Refactor the create_map function in the Dash app to utilize the new JavaScript function for enhanced map interactivity.
- Ensure proper namespace alignment between the Dash app and the external JavaScript file for seamless integration.
- Resolve JavaScript scope issues to address and fix the 'props not defined' error.
- Organize and optimize code structure for better readability and maintenance.
